requirements.txt
setuptools~=52.0.0
numpy~=1.21.2
opencv-python~=4.5.3.56
pyyaml~=5.1.1
